// src/main/webapp/app/shared/model/enums/email-template-type.enum.ts
export enum EmailTemplateTypeEnum {
    FORM_REQUEST_EMAIL = 'FORM_REQUEST_EMAIL',
    FORM_REQUEST_EMAIL_EDIT = 'FORM_REQUEST_EMAIL_EDIT',
    APPROVAL_EMAIL = 'APPROVAL_EMAIL',
    REJECTION_EMAIL = 'REJECTION_EMAIL',
    SUBMISSION_RECEIVED_EMAIL = 'SUBMISSION_RECEIVED_EMAIL',
    STATUS_EMAIL = 'STATUS_EMAIL',

    ACTIVATION_EMAIL = 'ACTIVATION_EMAIL',
    ACTIVATION_EMAIL_SUCCESS = 'ACTIVATION_EMAIL_SUCCESS',
    CREATION_EMAIL = 'CREATION_EMAIL',
    PASSWORD_RESET_EMAIL = 'PASSWORD_RESET_EMAIL'
}

export enum EmailTemplateDefaultsEnum {
    ACCOUNT_LOGIN = 'ACCOUNT_LOGIN',
    ACTIVATION_KEY = 'ACTIVATION_KEY',
    SUBMISSION_LINK = 'SUBMISSION_LINK',
    NOTE = 'NOTE',
    PASSWORD_RESET_KEY = 'PASSWORD_RESET_KEY',
    SUBMISSION_ID = 'SUBMISSION_ID',
    ORGANIZATION_ID = 'ORGANIZATION_ID',
    FORM_ID = 'FORM_ID'
}